feat(radial-menu): 新增自定义轮盘菜单功能
- 新增自定义轮盘菜单功能
- 在脚本 API 中新增了键盘和鼠标输入模拟功能
- 更新了删除按钮组件的参数结构，统一使用命名参数

+++ common/src/main/kotlin/moe/forpleuvoir/hiirosakura/functional/itemeditor/componentwrapper/ItemEnchantments.kt
@@ -193,7 +193,7 @@ fun ItemEnchantmentsComponentEditor(
                 Text(text = IGLang.remove)
             }.Column { (index, _) ->
                 DeleteButton(
-                    {
+                    confirmMessage = {
                         HSLang.deleteConfirm(
                             "${enchantments[index].value().description.string} ${
                                 Translatable(
@@ -203,7 +203,7 @@ fun ItemEnchantmentsComponentEditor(
                             }"
                         )
                     },
-                    { recompose() }
+                    recompose = { recompose() }
                 ) {
                     enchantments.removeAt(index)
                     levels.removeAt(index)
